% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_nonslot_funs.R
\name{user_utils}
\alias{user_utils}
\alias{self}
\title{Utilities for User-Defined Functions and Expressions}
\usage{
self()
}
\description{
These utility functions make it easier for the user to define functions
and expressions to store into a \code{VariableMetadata} object.
}
