% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package_vame.R
\docType{package}
\name{vame}
\alias{vame-package}
\alias{vame}
\title{vame: Variable Metadata}
\description{
Define value spaces and other metadata for variables.
}
\author{
\strong{Maintainer}: Joonas Miettinen \email{joonas.miettinen@cancer.fi} (\href{https://orcid.org/0000-0001-8624-6754}{ORCID})

}
