% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_S4.R
\docType{class}
\name{VariableMetadata-class}
\alias{VariableMetadata-class}
\alias{VariableMetadata}
\title{Variable Metadata}
\usage{
VariableMetadata(var_dt, var_set_dt, vame_list = NULL)
}
\arguments{
\item{var_dt}{\verb{[data.table]}

Contains information for individual variables. Must contain at a minimum
column \code{var_nm}.

See \strong{Details} and \strong{Features} for more information.}

\item{var_set_dt}{\verb{[data.table]}

Contains information for sets of variables --- e.g. a common value space.
Must contain at a minimum columns
\itemize{
\item \code{id} \verb{[integer, character]}: Identifies each set of variable names. E.g.
\code{c("my_set_01", "my_set_02")} or \code{1:2}.
\item \code{var_nm_set} \verb{[list]}: Each list element contains a character string
vector of variable names. e.g. \code{list(c("a", "b"))}.
}

See \strong{Details} and \strong{Features} for more information.}

\item{vame_list}{\verb{[NULL, list]} (default \code{NULL})

A list of metadata concerning the whole \code{VariableMetadata} object.
Its elements can be anything you want.
E.g. \code{vame_list = list(dataset_version = "1.0.0")}.}
}
\description{
Create and make use of a \code{VariableMetadata} object. It contains
\itemize{
\item \code{var_dt}: A hidden \code{data.table} containing metadata for variables,
\item \code{var_set_dt}: Hidden metadata for variable sets,
\item \code{vame_list}: A hidden list of metadata for the \code{VariableMetadata} object
itself,
\item A number of functions in S4 slots for making use of the hidden metadata
via e.g. \code{vm@var_assert} where \code{vm} is the \code{VariableMetadata} object.
}

A \code{VariableMetadata} object is created via calling the
\code{vame::VariableMetadata} function.
See section \strong{Features} for what you can do with \code{VariableMetadata}
objects.
}
