% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{VariableMetadata}
\alias{VariableMetadata}
\title{Variable Metadata}
\arguments{
\item{var_nm_dt}{\verb{[data.table]}

Contains information for individual variables.}

\item{var_nm_set_dt}{\verb{[data.table]}

Contains information for sets of variables --- e.g. a common value space.}
}
\description{
Create a VariableMetadata object.
}
