% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_vame.R
\docType{package}
\name{vame}
\alias{vame}
\title{vame: Variable Metadata}
\description{
\code{vame} makes it simpler to define and make use of metadata pertaining
to one more variables (e.g. a tabular dataset).

\href{https://github.com/FinnishCancerRegistry/vame/actions/workflows/R-CMD-check.yaml}{\if{html}{\figure{https://github.com/FinnishCancerRegistry/vame/actions/workflows/R-CMD-check.yaml/badge.svg}}}
}
\section{Recommended installation}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{devtools::install_github(
"FinnishCancerRegistry/vame",
ref = readline("enter latest tag on github: ")
)
}\if{html}{\out{</div>}}
}

\section{News}{

\subsection{News for version 0.1.0}{
\subsection{Object vame news:}{

First tagged version. See ?vame::VariableMetadata.
}

}
}

