% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_S4.R
\name{VariableMetadata}
\alias{VariableMetadata}
\title{Variable Metadata}
\usage{
VariableMetadata(var_dt, var_set_dt)
}
\arguments{
\item{var_dt}{\verb{[data.table]}

Contains information for individual variables. Must contain at a minimum
column \code{var_nm}.}

\item{var_set_dt}{\verb{[data.table]}

Contains information for sets of variables --- e.g. a common value space.
Must contain at a minimum columns
\itemize{
\item \code{id} \code{\link{character}}: Identifies each set of variable names. E.g.
\code{"my_set"}.
\item \code{var_nm_set} \verb{[list]}: Each list element contains a character string
vector of variable names. e.g. \code{list(c("a", "b"))}.
}}
}
\description{
Create a VariableMetadata object.
}
\section{Functions}{

\itemize{
\item \code{vame::VariableMetadata}: Use this function to create a new
VariableMetadata object.
}
}

\examples{
vd <- vame::VariableMetadata(
  var_dt = data.table::data.table(
    var_nm = c("a", "b", "c"),
    flavour = c("tasty", "rancid", "bitter")
  ),
  var_set_dt = data.table::data.table(
    id = "set_01",
    var_nm_set = list(c("a", "b"))
  )
)
vd@var_rename("a", "A")
stopifnot(identical(vd@var_meta_get("A", "flavour"), "tasty"))
}
